refactor: votes.tsx, voteDetail.tsx => move files to ko folder
