+++ package.json
@@ -34,8 +34,8 @@
   "devDependencies": {
     "@commitlint/cli": "^17.6.6",
     "@commitlint/config-conventional": "^17.6.6",
-    "@types/react-copy-to-clipboard": "^5.0.4",
     "@types/negotiator": "^0.6.1",
+    "@types/react-copy-to-clipboard": "^5.0.4",
     "husky": "^8.0.3",
     "negotiator": "^0.6.3",
     "prettier": "^2.8.8",
